requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
langchain_pinecone
python-dotenv
-e .